main: automatically detect new PG reports
Allow "linking" the Reports directory so that new .json files
inside are automatically detected and parsed.

This should probably become the main way to use PG vault, but
there's a risk it won't work on some browsers, so for now add
it as additional feature.

This definitely needs more testing